transform: factor out `dataflow::monotonic::is_monotonic`
In preparation of making the `is_monotonic` function stack
safe, do the following refactorings:

- Factor out the recursive `is_monotonic` function in the
  `datafalow::monotonic` module in a new top-level module
  `monotonic`.
- Wrap `is_monotonic` in a nullary struct `MonotonicFlag`.
- Rename `is_monotonic` to `apply`.
- Lift `optimize_dataflow_monotonic` in the `dataflow`
  namespace.
